webapi: Add setaltsig endpoint.
Allow setting or clearing the alternate public key that signs json
requests from the client. Save a copy of the request and signature in
the db.

This also adds another route to put a ticket in the database. Add a fee
status to indicate no fee address has been assigned yet.